bussiness logic level authorization

+++ src/server/Employee/resolver.js
@@ -1,18 +1,52 @@
+import { EmployeeRole, HrRole } from "@/configs/roleConfigs";
+import { setAuthTokenCookie } from "@/utils/authCookies";
 import { createJWT } from "@/utils/jwt";
 import bcrypt from "bcryptjs";
 
 const QueryResolvers = {
-  async getAllEmployees(_parent, _args, { Employee }) {
-    return await Employee.find({}).exec();
-  },
-  async getEmployeeLeaves(_parent, { id }, { Employee }) {
-    return await Employee.findById(id).exec();
+  /**
+   *
+   * Get all Employees data
+   * @returns Employee
+   * @Role "HR"
+   * @Permission "READ:EMPLOYEE"
+   */
+  async getAllEmployees(_parent, _args, { Employee, authUser }) {
+    try {
+      if (!authUser) {
+        throw new Error("Not logged in??, please login!");
+      }
+      if (authUser.roles.includes(HrRole)) {
+        if (authUser.permissions.includes("READ:EMPLOYEE")) {
+          return await Employee.find({}).exec();
+        } else throw new Error("No appropriate permission for your role");
+      } else throw new Error("Not authorized!");
+    } catch (error) {
+      return new Error(error.message);
+    }
   },
 };
 
 const MutationResolvers = {
-  async registerEmployee(_parent, { input }, { Employee, req, res }) {
+  /**
+   *
+   * Registers Employee
+   * @returns Employee
+   * @Role "HR"
+   * @Permission "READ:EMPLOYEE"
+   */
+  async registerEmployee(_parent, { input }, { Employee, authUser, res }) {
     try {
+      if (!authUser) {
+        throw new Error("Not logged in??, You need to login as Hr");
+      }
+      if (!authUser.roles.includes(HrRole)) {
+        throw new Error("Not authorized!");
+      }
+      if (!authUser.permissions.includes("CREATE:EMPLOYEE")) {
+        throw new Error("No appropriate permission for your role");
+      }
+
       const existingEmp = await Employee.findOne({
         email: input.email,
       }).exec();
@@ -31,19 +65,14 @@ const MutationResolvers = {
         password: passwordHash,
         lastLoginAt: time_now,
       });
-      if (newEmp) {
-        const jwtToken = createJWT(newEmp, time_now);
-        return { token: jwtToken };
-      } else {
-        return null;
-      }
+      return newEmp;
     } catch (error) {
       return new Error("registering Employee error: " + error.message);
     }
   },
 
   // login mutation
-  async loginEmployee(_parent, { input }, { Employee, req, res }) {
+  async loginEmployee(_parent, { input }, { Employee, res }) {
     try {
       const { email, password } = input;
       const employee = await Employee.findOne({ email }).exec();
@@ -55,8 +84,11 @@ const MutationResolvers = {
         throw new Error("Invalid password!");
       } else {
         const time_now = new Date().toString();
+        employee.lastLoginAt = time_now;
+        await employee.save();
         const jwtToken = createJWT(employee, time_now);
-        return { token: jwtToken };
+        setAuthTokenCookie(res, jwtToken);
+        return employee;
       }
     } catch (error) {
       return new Error("Error logging in Employee: " + error.message);
@@ -64,8 +96,26 @@ const MutationResolvers = {
   },
 
   // leave
-  async applyLeave(_parent, { date, id }, { Employee }) {
+  /**
+   *
+   * Apply for leave
+   * @returns Leave
+   * @Role "Employee"
+   * @Permission "UPDATE:ME"
+   */
+  async applyLeave(_parent, { date, id }, { Employee, authUser }) {
     try {
+      if (!authUser) {
+        throw new Error(
+          "Not logged in??, You need to login before applying for leave"
+        );
+      }
+      if (!authUser.roles.includes(EmployeeRole)) {
+        throw new Error("Not authorized!");
+      }
+      if (!authUser.permissions.includes("UPDATE:ME")) {
+        throw new Error("No appropriate permission for your role");
+      }
       const updatedLeave = await Employee.findByIdAndUpdate(
         id,
         {
@@ -87,9 +137,24 @@ const MutationResolvers = {
     }
   },
 
-  // delete employee
+  /**
+   *
+   * Delete Employee account
+   * @returns boolean
+   * @Role "HR"
+   * @Permission "DELETE:EMPLOYEE"
+   */
   async removeEmployee(_parent, { id }, { Employee }) {
     try {
+      if (!authUser) {
+        throw new Error("Not logged in??, please login!");
+      }
+      if (!authUser.roles.includes(HrRole)) {
+        throw new Error("Not authorized!");
+      }
+      if (!authUser.permissions.includes("DELETE:EMPLOYEE")) {
+        throw new Error("No appropriate permission for your role!");
+      }
       const deletedEmp = await Employee.findByIdAndRemove(id).exec();
       return Boolean(deletedEmp);
     } catch (error) {
